==> src/ocr/10-prepare_dataset_supervised_ocr_ner_ref.ipynb <==
# Dataset preparation -- OCR and NER reference

## Directories

import os.path

DATASET_PATH = os.path.abspath("../../dataset")
DATASET_PATH

sup_dir = os.path.join(DATASET_PATH, "supervised")

json_dir = os.path.join(sup_dir, "00-dataset_article_das22.pdf")

## Summary of incorrect chars in annotated content
```
1 Bottin1_1820 107 True
2 Bottin1_1820 201 True
3 Bottin1_1820 339 True
4 Bottin1_1820 589 True
Suspect char@007: ƒ (0x192 LATIN SMALL LETTER F WITH HOOK -- cat.: Ll)
Suspect char@018: ƒ (0x192 LATIN SMALL LETTER F WITH HOOK -- cat.: Ll)
5 Bottin1_1827 37 True
6 Bottin1_1827 117 True
7 Bottin1_1827 452 True
8 Bottin1_1827 475 True
9 Bottin1_1837 80 True
10 Bottin1_1837 114 True
11 Bottin1_1837 203 True
12 Bottin1_1837 345 True
13 Bottin1_1837 663 True
14 Bottin1_1837 667 True
Suspect char@013: ƒ (0x192 LATIN SMALL LETTER F WITH HOOK -- cat.: Ll)
Suspect char@029: ƒ (0x192 LATIN SMALL LETTER F WITH HOOK -- cat.: Ll)
15 Bottin3_1854a 72 True
16 Bottin3_1854a 74 True
17 Bottin3_1854a 200 True
18 Bottin3_1854a 238 True
21 Bottin3_1854a 892 True
22 Bottin3_1854a 1049 True
Suspect char@038: с (0x441 CYRILLIC SMALL LETTER ES -- cat.: Ll)
Suspect char@039: а (0x430 CYRILLIC SMALL LETTER A -- cat.: Ll)
Suspect char@041: р (0x440 CYRILLIC SMALL LETTER ER -- cat.: Ll)
Suspect char@042: о (0x43e CYRILLIC SMALL LETTER O -- cat.: Ll)
Suspect char@044: х (0x445 CYRILLIC SMALL LETTER HA -- cat.: Ll)
Suspect char@086: с (0x441 CYRILLIC SMALL LETTER ES -- cat.: Ll)
Suspect char@087: а (0x430 CYRILLIC SMALL LETTER A -- cat.: Ll)
Suspect char@089: р (0x440 CYRILLIC SMALL LETTER ER -- cat.: Ll)
Suspect char@090: о (0x43e CYRILLIC SMALL LETTER O -- cat.: Ll)
Suspect char@092: х (0x445 CYRILLIC SMALL LETTER HA -- cat.: Ll)
23 Cambon_almgene_1841 141 True
24 Cambon_almgene_1841 301 True
25 Cambon_almgene_1841 330 True
26 Cambon_almgene_1841 375 True
27 Cambon_almgene_1841 418 True
28 Cambon_almgene_1841 487 True
29 Deflandre_1828 278 True
30 Deflandre_1828 310 True
31 Deflandre_1828 881 True
32 Deflandre_1828 937 True
33 Deflandre_1829 505 True
34 Deflandre_1829 743 True
35 Deflandre_1829 949 True
36 Deflandre_1829 1026 True
37 Didot_1841a 162 True
38 Didot_1841a 183 True
39 Didot_1841a 206 True
40 Didot_1841a 316 True
41 Didot_1841a 500 True
Suspect char@034: р (0x440 CYRILLIC SMALL LETTER ER -- cat.: Ll)
Suspect char@035: о (0x43e CYRILLIC SMALL LETTER O -- cat.: Ll)
Suspect char@061: р (0x440 CYRILLIC SMALL LETTER ER -- cat.: Ll)
Suspect char@062: о (0x43e CYRILLIC SMALL LETTER O -- cat.: Ll)
42 Didot_1841a 542 True
Suspect char@009: С (0x421 CYRILLIC CAPITAL LETTER ES -- cat.: Lu)
Suspect char@020: С (0x421 CYRILLIC CAPITAL LETTER ES -- cat.: Lu)
43 Didot_1851a 92 True
44 Didot_1851a 169 True
45 Didot_1851a 226 True
46 Didot_1851a 415 True
47 Didot_1851a 419 True
48 Didot_1851a 639 True
49 Didot_1851a 698 True
52 Didot_1854a 83 True
53 Didot_1854a 326 True
Suspect char@112: ٭ (0x66d ARABIC FIVE POINTED STAR -- cat.: Po)
54 Didot_1854a 607 True
58 DidotBottin_1860a 186 True
59 DidotBottin_1860a 280 True
60 DidotBottin_1861a 238 True
61 DidotBottin_1861a 424 True
62 Duverneuil_et_La_Tynna_1801 260 True
63 Duverneuil_et_La_Tynna_1801 371 True
64 Duverneuil_et_La_Tynna_1801 401 True
65 Duverneuil_et_La_Tynna_1801 415 True
66 Duverneuil_et_La_Tynna_1801 454 True
67 Duverneuil_et_La_Tynna_1805 193 True
68 Duverneuil_et_La_Tynna_1805 250 True
Suspect char@004: ˘ (0x2d8 BREVE -- cat.: Sk)
Suspect char@032: ˘ (0x2d8 BREVE -- cat.: Sk)
69 Duverneuil_et_La_Tynna_1805 251 True
70 Duverneuil_et_La_Tynna_1805 292 True
71 Duverneuil_et_La_Tynna_1805 305 True
72 Duverneuil_et_La_Tynna_1806 147 True
73 Duverneuil_et_La_Tynna_1806 220 True
76 Favre_et_Duchesne_1798 375 True
77 Favre_et_Duchesne_1798 429 True
78 Favre_et_Duchesne_1798 625 True
79 Favre_et_Duchesne_1798 658 True
80 Favre_et_Duchesne_1798 700 True
81 Favre_et_Duchesne_1798 701 True
Suspect char@010: В (0x412 CYRILLIC CAPITAL LETTER VE -- cat.: Lu)
Suspect char@015: В (0x412 CYRILLIC CAPITAL LETTER VE -- cat.: Lu)
82 La_Tynna_1813 163 True
83 La_Tynna_1813 166 True
84 La_Tynna_1813 346 True
85 La_Tynna_1813 377 True
86 notables_communaux_seine_1801 57 True
87 notables_communaux_seine_1801 144 True
```

## Fix bad chars before reimport
This is idempotent and was run before relaunching the notebook.

# uncorrected OCR char
!parallel sed -i s/ƒ/f/g ::: $json_dir/0004.json $json_dir/0014.json

# wrong detection, this is a capital B over black circle
!sed -i s/Ⓡ/Ⓑ/g $json_dir/0021.json

# mixed scripts (cyrillic and latin) in OCR output
!sed -i s/Quinсаmроiх/Quincampoix/g $json_dir/0022.json

# mixed scripts (cyrillic and latin) in OCR output
!sed -i s/роix/poix/g $json_dir/0041.json

# mixed scripts (cyrillic and latin) in OCR output
!sed -i 's/С. F. 1839./C. F. 1839./g'  $json_dir/0042.json

# wrong chars
!sed -i 's/★/*/g;s/٭/*/g' $json_dir/0053.json

# noise (Sk diacritics)
!sed -i "s/˘//g"  $json_dir/0068.json

# mixed scripts (cyrillic and latin) in OCR output
!sed -i 's/Mallet (J.В.)/Mallet (J.B.)/g' $json_dir/0081.json

## Build file list

import pandas as pd

annot_file = os.path.join(sup_dir, "annotation_table.csv")
annot_file

!head $annot_file

annot_metadata = pd.read_csv(annot_file)
annot_metadata

import json
from text_utils import (
    simplify_unicode_charset, 
    replace_annotation_codes, 
    fix_manual_ner_xml, 
    check_alignment_charset, 
    remove_xml_tags_and_entities)

## Gather and normalize all NER and OCR references
We remove empty tags.
We drop inconsistent entries (OCR text and NER XML).

dataset = []
all_raw_ocr_txt = []
all_new_ocr_txt = []
limit = 10000000
count = 0
lost_entries = 0
bad_boxes = 0
for item_id, book, page, in_dataset in zip(
        annot_metadata["#"], annot_metadata["PDF"], annot_metadata["VUE"], annot_metadata["Inclus dataset"]):
    print("="*30)
    print(item_id, book, page, in_dataset)
    if not in_dataset:
        print("\t >> SKIPPING (not in dataset)")
        continue
    content = None
    with open(os.path.join(json_dir, f"{item_id:04d}.json")) as infile:
        content = json.load(infile)
    for elem in content:
        if count > limit:
            break
        if elem["type"] != "ENTRY":
            continue
        if "checked" not in elem or not elem["checked"]:
            continue
        count += 1

        newelem = {}
        for k in ["id", "box"]:
            newelem[k] = elem[k]
        newelem["book"] = book
        newelem["page"] = page

        entry_id = elem["id"]

        x, y, w, h = elem["box"]
        valid_box = x >= 0 and y >= 0 and w > 10 and h > 10
        newelem["valid_box"] = valid_box
        if not valid_box:
            bad_boxes += 1
            print(f"ERROR: ({entry_id}) bad box: {x:.0f}, {y:.0f}, {w:.0f}, {h:.0f}")
            # We keep these entries for now, as NER can be valid
            # We tag them for later filtering (for OCR eval)

        orig_ocr_txt = elem["text_ocr"].strip()
        new_ocr_txt = simplify_unicode_charset(replace_annotation_codes(orig_ocr_txt))
        newelem["text_ocr"] = new_ocr_txt
        if not check_alignment_charset(new_ocr_txt):
            print(f"FATAL: ({entry_id}) Bad charset")
            raise RuntimeError("Should not have bad charset after normalization.")
        
        orig_ner_xml = elem["ner_xml"].strip()
        new_ner_xml = fix_manual_ner_xml(orig_ner_xml)
        newelem["ner_xml"] = new_ner_xml

        orig_ner_text = remove_xml_tags_and_entities(orig_ner_xml)
        if not orig_ner_text == orig_ocr_txt:
            print(f"ERROR: ({entry_id}) NER and OCR don't match.")
            print(".....: OCR: \t%s"%(orig_ocr_txt.replace('\n', '\u2424')))
            print(".....: NER: \t%s"%(orig_ner_text.replace('\n', '\u2424')))
            lost_entries += 1
            continue

        dataset.append(newelem)
        all_raw_ocr_txt.append(orig_ocr_txt)
        all_new_ocr_txt.append(new_ocr_txt)
        

lost_entries

bad_boxes

len(dataset)

dataset[:3]

## Export

out_path = os.path.join(sup_dir, "10-ref-ocr-ner-json", "all.json")
out_path

with open(out_path, 'w') as out_file:
    json.dump(dataset, out_file, indent=2)

!head $out_path

out_path_csv = os.path.join(sup_dir, "10-ref-ocr-ner-json", "gold.csv")

lines_to_write = []
for entry in dataset:
    line = '"' + entry['ner_xml'] + '", "' + entry['book'] + '"'
    lines_to_write.append(line)
all_lines = "\n".join(lines_to_write)

import io

with io.open(out_path_csv, "wt", encoding="UTF-8", newline='',
            errors="strict") as file_output:
    file_output.write(all_lines)

!head $out_path_csv

**WARNING** we have much more escaped XML entities now than in previous gold version.
