==> src/ocr/80-prepare_dataset_ocr_text_files.ipynb <==
# Prepare text files for each entry
- reference
- Pero OCR
- Tesseract OCR

import os.path

DATASET_PATH = os.path.abspath("../../dataset/supervised")
DATASET_PATH

ref_path = os.path.join(DATASET_PATH, "10-ref-ocr-ner-json", "all.json")
pero_path = os.path.join(DATASET_PATH, "21-ocr-pero-final", "all.json")
tess_path = os.path.join(DATASET_PATH, "22-ocr-tess-final", "all.json")

## Text file generation

import json
import io

def generate_text_files(json_path, output_dir):
    entries = None
    with open(json_path) as in_file:
        entries = json.load(in_file)
    for entry in entries:
        out_filename = f'{entry["book"]}-{entry["page"]:04d}-{entry["id"]:04d}.txt'
        out_path = os.path.join(output_dir, out_filename)
        text = entry["text_ocr"]
        with io.open(out_path, "wt", encoding="UTF-8", newline='', errors="strict") as file_output:
            file_output.write(text)


for (in_, out) in [(ref_path, "ref"), (pero_path, "pero"), (tess_path, "tess")]:
    out_dir = os.path.join(DATASET_PATH, "80-ocr-text-files", out)
    print(out)
    generate_text_files(in_, out_dir)
